# static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Example API for requesting and return products requests",
        "version": "1.0.0",
        "title": "Products API",
        "contact": {
            "email": "fer13vlc@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "tags": [
        {
            "name": "Products",
            "description": "Example API for requesting and return products requests"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Returns Products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/products"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "requestBody": {
                    "description": "Product Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/products"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/products"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/products/{name}": {
            "parameters": [
                {
                    "name": "name",
                    "in": "path",
                    "required": true,
                    "description": "Product name",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products with given name",
                "parameters": [
                    {
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "description": "Product name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/products"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    },
                    "404": {
                        "description": "Failed. Book request not found."
                    }
                }
            },
            "put": {
                "summary": "edit a product by name",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "description": "Product name"
                    }
                ],
                "requestBody": {
                    "description": "Product Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/products"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/products"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            },
            "delete": {
                "summary": "Delete Products by Name",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "description": "Product name",
                        "schema": {
                            "$ref": "#/components/schemas/products"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/products"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    },
                    "404": {
                        "description": "Failed. Book Request not found."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "products": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "format": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "stock": {
                        "type": "number"
                    }
                }
            }
        }
    }
}